layer bootstrap sampling atop simulation
In addition to simulating missing artwork data with n iterations, also samples from each iteration/year combination m times, resulting in n*m calculations of diversity per year. The median and upper/lower bounds are then summarized and plotted.

As a result of n*m complexity, I've drastically reduced the number of iterations for each round that the user is allowed to input.